# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Hotel Las Vegas Reservation Management System API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /rooms:
    get:
      operationId: listRooms
      tags: [rooms]
      summary: List all rooms
      parameters:
        - name: status
          in: query
          required: false
          schema:
            type: string
        - name: type
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: List of rooms
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Room"
    post:
      operationId: createRoom
      tags: [rooms]
      summary: Create a room
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RoomInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Room"
  /rooms/{id}:
    get:
      operationId: getRoom
      tags: [rooms]
      summary: Get a room
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Room
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Room"
        "404":
          description: Not found
    patch:
      operationId: updateRoom
      tags: [rooms]
      summary: Update a room
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RoomUpdate"
      responses:
        "200":
          description: Updated room
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Room"
        "404":
          description: Not found
    delete:
      operationId: deleteRoom
      tags: [rooms]
      summary: Delete a room
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
...
components:
  schemas:
...
    Room:
      type: object
      required: [id, number, type, floor, status, pricePerNight, createdAt]
      properties:
        id:
          type: integer
        number:
          type: string
        type:
          type: string
        floor:
          type: integer
        status:
          type: string
        pricePerNight:
          type: number
        description:
          type: ["string", "null"]
        createdAt:
          type: string
    RoomInput:
      type: object
      required: [number, type, floor, pricePerNight]
      properties:
        number:
          type: string
        type:
          type: string
        floor:
          type: integer
        pricePerNight:
          type: number
        description:
          type: string
    RoomUpdate:
      type: object
      properties:
        number:
          type: string
        type:
          type: string
        floor:
          type: integer
        status:
          type: string
        pricePerNight:
          type: number
        description:
          type: string
...